refactor(backend): fold chat-input guard into retrieval/safety.py
Keep the oversized-input guard with the other agent safety guards
(AgentSafetyGuard) instead of a standalone module. This avoids growing
the grandfathered backend/utils/retrieval package file count past its
architecture-map ratchet baseline, and safety.py is already import-light
so the pure-logic unit tests stay free of the heavy chat/LLM stack.

No behavior change: same fit_within_budget / message_text / constants,
same wiring in execute_agentic_chat_stream.
